Move RuleConfigTrait to rule.rs
